[Codegen][TopK] Faster TopK implementation
This is a hybrid implementation of TopK.
The load and compare of elements is done using vectors and the
sorting/writing to result is done using scalar registers. The current
implementation is pure scalar.

This results in significant speedup of the function. The speedup scales
with the reduction of the ratio of output/input sizes.

This implementation seems to be about 10% faster even on smaller data
sets where the output/input ration is 1 (see below that benchmark for
input 1x32->1x32, which is basically using TopK for sorting).

For bigger data sets we observe even bigger speedups:
1. For 2x1024 -> 2x40 the speedup is 9.59x
2. For 1x256128 -> 2x40 the speedup is 5.46x.