[hip][cuda] Update event allocation and collection.
The existing system was not sufficient for graphs,
as they can be run out of order and have different
behavior for event recording.

This does not entirely solve the problem for re-use
if we ever want to simultaneously submit more than one graph
at a time. However this can be extended to do so in
the future.

The original issue was that graphs do not put their events
into the waiting state until they are actually dispatched.
This means that we were routinely trying to poll events
before they were ever submitted to the GPU. This caused
graphs to provide incorrect results in the resulting trace.
Furthermore sometimes graphs were submitted out of order,
this caused events to be returned out of order which
caused tracy to fail.

The broad solution is to no longer have an ordered circular
array of events and instead have a pool of events and
track their lifetime with more granularity so that we can
determine which ones are created, submitted and retired
with more accuracy.

Signed-off-by: Andrew Woloszyn <andrew.woloszyn@gmail.com>