[Codegen][GPU] Add kernel config for LLVMGPUTileAndFuse
This adds kernel configuration logic for targeting simple thread
distribution of linalg-based dispatches on LLVMGPU. The configuration
logic is primarily copied from the same logic on the SPIR-V side due to
the already well tested heuristics there for the kinds of varied target
descriptions that are present on the SPIR-V side.

Currently this is locked behind a flag
`iree-codegen-llvmgpu-use-tile-and-fuse`. Future patches will add
specialized logic for matmul.