[Flow] Make sink reshapes changes less conservative.
While deciding if a reshape needs "sinking", for a
`tensor.expand_shape` -> `linalg.*`, first check was to check that the
`linalg.*` operation could already fuse with one of its existing
producers. That check was broadly aggressive. The fusion only kicks in
when the iteration domains match. Eventually the actual dispatch
formation logic needs to be commoned to a single place to do this
better, but kicking that to a follow up.

Signed-off-by: MaheshRavishankar <mahesh.ravishankar@gmail.com>